% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addPctColumn}
\alias{addPctColumn}
\title{Add a "pctLABEL" column that shows the percent of total for that label (EMP) in that NAICS code}
\usage{
addPctColumn(data, columnLabels)
}
\arguments{
\item{data}{tibble}

\item{columnLabels}{character[n]}
}
\value{
Tibble with "pctCOLUMNLABEL" column attached
}
\description{
Add a "pctLABEL" column that shows the percent of total for that label (EMP) in that NAICS code
}
