% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addScoreColumn}
\alias{addScoreColumn}
\title{Add a "score" column that calcs score based on both industry and label weights}
\usage{
addScoreColumn(data, industryWeights, metricWeights)
}
\arguments{
\item{data}{tibble}

\item{industryWeights}{data.frame with 'labels' and 'weights' columns}

\item{metricWeights}{data.frame with 'labels' and 'weights' columns}
}
\value{
Tibble with "score" column attached
}
\description{
Add a "score" column that calcs score based on both industry and label weights
}
