% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aggregateStateScores}
\alias{aggregateStateScores}
\title{Convert a tibble with "stateName" and "totalScore" column into a simple data.frame for plotting}
\usage{
aggregateStateScores(data)
}
\arguments{
\item{data}{tibble}
}
\value{
Tibble with state, totalScore, long, lat, and printScore cols
}
\description{
Convert a tibble with "stateName" and "totalScore" column into a simple data.frame for plotting
}
