% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{censusData}
\alias{censusData}
\title{Census Bureau data set}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22238 rows and 18 columns.}
\usage{
censusData
}
\description{
Contains state and NAICS detail from the Census Bureau
}
\keyword{datasets}
