% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data2stateScores}
\alias{data2stateScores}
\title{Convert weight inputs and raw data into aggregated scores by state}
\usage{
data2stateScores(rawData, metricWeights, industryWeights)
}
\arguments{
\item{rawData}{tibble representing Census Bureau data set}

\item{metricWeights}{data.frame with "labels" (char.) and "weights" (num.)}

\item{industryWeights}{data.frame with "labels" (char.) and "weights" (num.)}
}
\value{
A tibble with state, totalScore, long, lat, and printScore cols
}
\description{
Convert weight inputs and raw data into aggregated scores by state
}
