% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filterIndustries}
\alias{filterIndustries}
\title{Filter a tibble by industry}
\usage{
filterIndustries(dataToFilter, industries)
}
\arguments{
\item{dataToFilter}{tibble}

\item{industries}{character[n]. Industry codes to include}
}
\value{
tibble with only the NAICS codes in industries
}
\description{
Filter a tibble by industry
}
