% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getTotal}
\alias{getTotal}
\title{Calculate the total of a column, aggregated by NAICS.id}
\usage{
getTotal(data, naics, columnLabel)
}
\arguments{
\item{data}{tibble}

\item{naics}{vector of NAICS.id values}

\item{columnLabel}{expr[1]. Column heading - ESTAB, EMP, etc.}
}
\description{
Calculate the total of a column, aggregated by NAICS.id
}
