% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makeWeightsInput}
\alias{makeWeightsInput}
\title{Create paired weight and label input objects}
\usage{
makeWeightsInput(labels, weights)
}
\arguments{
\item{labels}{character[n].}

\item{weights}{numeric[n].}
}
\value{
data.frame with labels and weights columns
}
\description{
Create paired weight and label input objects
}
