% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotDemandMap}
\alias{plotDemandMap}
\title{Create a ggplot object representing the demand map}
\usage{
plotDemandMap(mapData, stateScoresData)
}
\arguments{
\item{mapData}{data.frame w/ long, lat, and region columns; region = state name. This provides the lat/lon for the state boundaries.}

\item{stateScoresData}{tibble w/ state, totalScore, printScore, lat, and long. This provides the lat/lon for the state midpoints.}
}
\value{
ggplot
}
\description{
Create a ggplot object representing the demand map
}
