% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotRegionMap}
\alias{plotRegionMap}
\title{Create a ggplot object representing a sales territory map}
\usage{
plotRegionMap(mapData, regionData)
}
\arguments{
\item{mapData}{data.frame w/ long, lat, and region columns; region = state name. This provides the lat/lon for the state boundaries.}

\item{regionData}{tibble w/ state and salesRegion columns}
}
\value{
ggplot
}
\description{
Create a ggplot object representing a sales territory map
}
