% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{removeMapAxisLabels}
\alias{removeMapAxisLabels}
\title{Remove axis labels and ticks from a ggplot object}
\usage{
removeMapAxisLabels(ggplotObject)
}
\arguments{
\item{ggplotObject}{ggplot}
}
\value{
ggplot
}
\description{
Remove axis labels and ticks from a ggplot object
}
